SocKult_RumDet/requirements.txt
pandas
numpy
nltk
pytorch
transformers
sklearn
gensim
keras
tensorflow